fix: model labels as usize instead i64

+++ dynamic_learned_index/src/index.rs
@@ -386,7 +386,7 @@ impl LevelIndexBuilder {
             .device(self.model_device.to_tch_device())
             .input_nodes(input_shape as i64)
             .train_params(model_config.train_params.clone())
-            .labels(n_buckets as i64)
+            .labels(n_buckets)
             .label_method(distance_fn.clone().into());
         model_config.layers.iter().for_each(|layer| {
             model_builder.add_layer(*layer);
